fix(php-8.5+forms): SupplierType finishView + repository scalar casts
Three production runtime bugs after PHP 8.5 base-image bump:

* `BCExerciseRepository::getStatistics()` line 88 — `round($avgSuccessRating
  ?? 0, 2)` exploded with "Argument #1 ($num) must be of type int|float,
  string given" because Doctrine `getSingleScalarResult()` returns string
  for AVG/SUM/COUNT on MySQL/PDO. Cast to (float) before round.
* `InterestedPartyRepository` (same root cause), `avg_satisfaction` field
  rounded same way. Same (float) cast.
* `ComplianceRequirementRepository::getFrameworkStatisticsForTenant()` —
  ALL four COUNT scalars now cast to (int) at the source. Fixes 6
  downstream consumers (ComplianceExtension Twig filter, 3
  ComplianceAssessmentService methods, ComplianceFrameworkRepository,
  ComplianceFrameworkController) without local patches.

Plus form bug:

* `SupplierType::buildView()` — accessing `$view['subcontractorChain']`
  inside buildView is too early; child FormViews are populated AFTER
  buildView returns. Renamed to `finishView()` (correct hook).
  `processingLocations` array-to-string-conversion warning hardened with
  explicit `is_scalar` filter for legacy nested-data resilience.

Audit covered all 96 src/-tree files / 344 round() occurrences. The
remaining ~20 risky string-coercion locations identified are in
non-critical paths and are being patched in subsequent commits.

+++ src/Repository/BCExerciseRepository.php
@@ -85,7 +85,7 @@ class BCExerciseRepository extends ServiceEntityRepository
             'total' => $total,
             'completed' => $completed,
             'planned' => $planned,
-            'avg_success_rating' => round($avgSuccessRating ?? 0, 2)
+            'avg_success_rating' => round((float) ($avgSuccessRating ?? 0), 2)
         ];
     }
 }
